Infinite voting fix.
If there is an wrong code in the plugin and the sourcemod interrupts the execution of the code or the plugin author forgot to add DisplayVoteFail or DisplayVotePass, then the voting table will be endlessly shown to the players. Now the extension will do it on its own and send the players that the vote failed. The fix is primarily made for the game left4dead2, i don't have any other games to test the change, but this code probably won't have any problems in other games.